refactor: extract rolling metric output assembly
