+++ quality/baseline_report.md
@@ -10,7 +10,7 @@ documentation posture, and validation gates. It is not a completion claim.
 
 ## Code Size Baseline
 
-- Python source files under `src/`: 188
+- Python source files under `src/`: 189
 - Python test files under `tests/`: 93
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -38,7 +38,7 @@ documentation posture, and validation gates. It is not a completion claim.
 | src/app/services/concentration/parsing.py | 319 | 11074 |
 | tests/integration/test_risk_calculate_live_characterization.py | 314 | 12945 |
 | tests/integration/test_historical_attribution_live_characterization.py | 305 | 12343 |
-| src/app/services/rolling_metric_series.py | 299 | 10413 |
+| src/app/services/attribution_decomposition.py | 295 | 9686 |
 
 ### Largest Functions And Classes
 
@@ -89,12 +89,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-295 files already formatted
+296 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 281 source files
+Success: no issues found in 282 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -108,10 +108,10 @@ Success: no issues found in 281 source files
 ...
 src\app\upstream_errors.py                                    55      0     20      3    96%   181->194, 187->189, 192->194
 ------------------------------------------------------------------------------------------------------
-TOTAL                                                       5208     80    978     79    97%
+TOTAL                                                       5212     80    978     79    97%
 
 148 files skipped due to complete coverage.
-474 passed in 8.29s
+474 passed in 7.81s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -141,7 +141,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 188 files...
+Scanning 189 files...
 
 Success! No dependency issues found.
 ```
@@ -240,7 +240,7 @@ tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matri
 tests/unit/test_upstream_errors.py::test_invalid_payload_and_missing_data_carry_structured_categories
 tests/unit/test_upstream_errors.py::test_extract_upstream_error_detail_variants
 
-474 tests collected in 1.60s
+474 tests collected in 1.58s
 ```
 - Import-linter report-only: passed
